refactor: rewrite fmt_packet
Normally I would implement Display but Packet is from another crate so that doesnt work.
Second best is writing to a String with the write macro.

This code should be equivalent to yours, but it seems not correct like the payload only being used when there is a content format. It could be always used?